perf:处理删除文件后，创建新文件时失败，丢失原文件数据的问题

+++ utils/useUpdateFile.js
@@ -1,11 +1,15 @@
 import dayjs from 'dayjs'
-import { planPrev, separator } from '../config/index'
+import {
+  planPrev,
+  separator
+} from '../config/index'
 import {
   setFileName,
   readFile,
   removeFile,
   createLocalFile,
-  verifyDataFormat
+  verifyDataFormat,
+  renameFile
 } from './util'
 
 /**
@@ -29,13 +33,33 @@ const useUpdateFile = (oldVal, newVal) => {
 
   const newText = text.replace(JSON.stringify(oldVal), JSON.stringify(newVal))
 
-  const isDelete = removeFile(oldFileName)
-  if (!isDelete) {
-    return false
+  // 将要删除的文件，先重命名，避免创建新文件失败，数据丢失
+  const oldFileName_new = 'delete_' + oldFileName
+  const isRenameDel = renameFile(oldFileName, oldFileName_new)
+
+  // 重命名失败，直接删除
+  if (!isRenameDel) {
+    const isDelete = removeFile(oldFileName)
+    if (!isDelete) {
+      return false
+    }
   }
 
   const newFileName = setFileName(planPrev, dayjs(newVal.date).format('YYYYMM'))
   const isSuccess = createLocalFile(newText, newFileName)
+
+  // 创建新文件失败，恢复原文件
+  if (!isSuccess && isRenameDel) {
+    renameFile(oldFileName_new, oldFileName)
+  }
+
+  // 创建新文件成功，删除原文件
+  if (isRenameDel) {
+    setTimeout(() => {
+      removeFile(oldFileName_new)
+    }, 0)
+  }
+
   return isSuccess
 }
